+++ NEWS.md
@@ -1,3 +1,10 @@
+# gfilmm 2.0.0
+
+* The main function `gfilmm` now has an option `nthreads`, to run some pieces of code of the algorithm in parallel. 
+* Use of a sparse matrix to reduce memory consumption and to improve the speed.
+* New function `gfiQuantile`.
+
+
 # gfilmm 1.0.1
 
 * Removed the quadruple precision feature because of issues with the CRAN checks.